service/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
wrapt==1.14.0
